test: remove redundant tests from catch-all and relocate unique ones
Deleted ~20 tests from test_remaining_branches.py that were exact
duplicates of tests in their dedicated _extended files. Moved the two
genuinely unique tests (get_state exception handler and dict-under-
sensitive-key sanitize path) to test_heating_extended.py and
test_helpers.py respectively. Coverage remains at 100%.
